train: Fix synth

+++ train.py
@@ -209,7 +209,7 @@ class Trainer:
                     # [1, T]
                     synth, _ = self.model.forward(
                         torch.tensor(speech[None], device=self.wrapper.device),
-                        torch.tensor(pitch[None], device=self.wrapper.device))
+                        torch.tensor(pitch[None], device=self.wrapper.device, dtype=torch.float32))
 
                     synth = synth.squeeze(dim=0).cpu().numpy()
                     self.test_log.add_image(
